+++ docgen/libraries.go
@@ -12,6 +12,7 @@ import (
 	"github.com/jig/lisp/lib/sql/nssql"
 	"github.com/jig/lisp/lib/system/nssystem"
 	"github.com/jig/lisp/lib/test/nstest"
+	"github.com/jig/lisp/lib/web/nsweb"
 	"github.com/jig/lisp/types"
 )
 
@@ -40,6 +41,7 @@ func standardLibraries() []library {
 		{"sql", nssql.Load},
 		{"cli", nscli.Load},
 		{"integrity", nsintegrity.Load},
+		{"web", nsweb.Load},
 		{"test", nstest.Load},
 	}
 }